Update the method for determining kernel language
The previous algorithm for `getNotebookType()` was based on getting the
kernel spec out of the current notebook. This is a problem when format
is triggered by, for example, `jupyterlab-autosave-on-focus-change`,
immediately after notebook creation. In that case, the language is
determined to be `null` and errors arise.

Now the `onSave()` handler waits for the `sessionContext` to be ready,
then `getNotebookType()` gets the same kernel spec info from the session
rather than the loaded notebook.